Fixed issue where only one apple would be collected if two were collided with at the same time.
Switched from using a groupcollide in the game loop to using a spritecollide in the apple's update function. I had to use a kill function to remove the apple rather than using dokilla and dokillb which are included in the groupcollide function. I also needed to include a statement at the end of the mapping loop which killed all the sprited used in the map drawing phase so that those apples wouldn't be counted when collided with