Starting to implement vision impairement
Stopped drawing all sprites naturally and now only draw the player, bricks, doors and windows. Apples and bricks will be added to the draw_sprites_group when they are within 150px of the player and then they will be drawn. Had issue where apples continued to be drawn after being collected and fixed this with the death variable. Also it took a little while to get the if statement correct. Still need to implement not being able to see through doors and bricks.