Error Fix
If an apple or player was not created during the map drawing process the game would tell the player to place one but when the player moved back to drawing their progress was lost. I fixed this by putting the loops for no_player and no_apple inside the map_draw loop.