Fixed
Check function still has issues but the game now runs correctly on second attempt. Previously the is_player_there variable remained true into the second iteration of map drawing which meant a player could not be placed. I have now set it to be false whenever the game loops so that the map drawing can be played over and over as intended