added deepcopy
previously I had been unaware that using a "=" sign would just create another name which references the same address and, due to python's passing method, I was updating my map within the check method which was causing the player sprite to be erased and hence ending the game immediately. I replaced the map in the check function with a deepcopy of the map called dummy_map and this solved the issue. It is not fully tested yet but the check seems to be working correctly