check working
check now works but can take a while on invalid and complex maps. Needed to implement an extra algorithm which searches randomly rather than based on the position of the apple. This will run if the original algorithm doesn't find all the apples in 10000 iterations and takes longer to find apples but does work.